Bug 1842067 - Filter by OriginAttributesPattern when purging bounce tracker site data. r=bvandersloot
Differential Revision: https://phabricator.services.mozilla.com/D223752

UltraBlame original commit: 69105a17312697f5c31311105282d2b5c8f76e6e

+++ toolkit/components/antitracking/bouncetrackingprotection/BounceTrackingProtection.cpp
@@ -1077,14 +1077,35 @@ nsresult BounceTrackingProtection::PurgeBounceTrackersForStateGlobal(
       cb->OnDataDeleted(0);
     } else {
       
-
-      
       
       nsAutoCString hostToPurge(host);
       nsContentUtils::MaybeFixIPv6Host(hostToPurge);
 
+      
+      
+      
+      
+      
+      
+      const OriginAttributes& oa = aStateGlobal->OriginAttributesRef();
+
+      nsAutoString oaPatternString;
+      OriginAttributesPattern pattern;
+      
+      
+      
+      pattern.mUserContextId.Construct(oa.mUserContextId);
+      pattern.mPrivateBrowsingId.Construct((oa.mPrivateBrowsingId));
+      pattern.mGeckoViewSessionContextId.Construct(
+          oa.mGeckoViewSessionContextId);
+
+      if (NS_WARN_IF(!pattern.ToJSON(oaPatternString))) {
+        
+        continue;
+      }
+
       rv = clearDataService->DeleteDataFromSiteAndOriginAttributesPatternString(
-          hostToPurge, u""_ns, false, TRACKER_PURGE_FLAGS, cb);
+          hostToPurge, oaPatternString, false, TRACKER_PURGE_FLAGS, cb);
       if (NS_WARN_IF(NS_FAILED(rv))) {
         clearPromise->Reject(0, __func__);
       }
